% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesQVAR-package.R
\docType{package}
\name{bayesQVAR-package}
\alias{bayesQVAR}
\alias{bayesQVAR-package}
\title{bayesQVAR: Bayesian Estimation of Quantile Vector Autoregression and Revelant Analysis Tools}
\description{
Bayesian estimation, quantile forecast and impulse response analysis of Quantile Vector Autoregression(QVAR) model.
}
\author{
\strong{Maintainer}: Fengyi Zhu \email{zhufengyi810@live.com}

}
\keyword{internal}
