% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesQVAR_class.R
\name{BQVAR}
\alias{BQVAR}
\title{BQVAR class}
\usage{
BQVAR(x)
}
\arguments{
\item{x}{List that contains estimation results of Bayesian quantile vector autoregression (BQVAR).}
}
\value{
\code{data_end}: Data.frame for the endogenous variables.

\code{data_exo}: Data.frame for the exogenous variables.

\code{lag}: Integer for the lag order.

\code{alpha}: Double for the quantile level.

\code{prior}: List for the prior settings.

\code{samplerSetting}: List for the sampler settings.

\code{method}: Character string for the estimation method.

\code{estimates}: List for the estimation results.

\code{mcmcChains}: List for the MCMC chains.

\code{residuals}: List for the residuals.
}
\description{
Class for Bayesian quantile vector autoregression (BQVAR) model, which is only used to create an object of formally defined class when return the output of \link{estBQVAR}.
}
