% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesQVAR_class.R
\name{BQVAR_forecast}
\alias{BQVAR_forecast}
\title{BQVAR_forecast class}
\usage{
BQVAR_forecast(x)
}
\arguments{
\item{x}{List that contains quantile forecast.}
}
\value{
An object of class \code{BQVAR_forecast}, which contains:

\code{modelSpecif}: ModelSpecif list for the model specification.

\code{horizon}: Integer for the forecast horizon.

\code{probPath}: Matrix for the quantile path.

\code{forecastList}: List for the forecasts of quantile and mean. quantile.
}
\description{
convert a list of quantile forecasts into a formally defined S7 class,which is only used to create an object of formally defined class when return the output of \link{forecastQuant}.
}
