% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesQVAR_class.R
\name{BQVAR_QIRF}
\alias{BQVAR_QIRF}
\title{BQVAR_QIRF class}
\usage{
BQVAR_QIRF(x)
}
\arguments{
\item{x}{List that contains the quantile impulse response functions and its' confidence intervals.}
}
\value{
An object of class \code{BQVAR_QIRF}, which contains:

\code{modelSpecif}: List for the model specification.

\code{horizon}: Integer for the forecast horizon.

\code{probPath}: Matrix for the quantile path.

\code{names_impulse}: Character vector for the impulse variable names.

\code{names_response}: character vector for the response variable names.

\code{shockScale}: Numeric vector for the shock scale of each variable.

\code{confInt}: Logical for whether the confidence intervals are estimated.

\code{credInt}: Logical for whether the credibility intervals are estimated.

\code{alpha}: Numeric vector for the significant level.

\code{qirfList}: List for the quantile impulse response functions and its' confidence intervals.

\code{confIntList}: List of the confidence intervals.

\code{credIntList}: List of the credibility intervals.
}
\description{
Class for quantile forecasts based on QVAR, which is only used to create an object of formally defined class when return the output of \link{calQIRF}.
}
