% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuData.R
\name{simuData}
\alias{simuData}
\title{Simulate QVAR(1)}
\usage{
simuData(n, alpha, sampleSize)
}
\arguments{
\item{n}{Dimension of the QVAR(1) process.}

\item{alpha}{Vector of the quantile level of interest, \eqn{\bm{\alpha}^*},.}

\item{sampleSize}{The sample size of simulated data.}
}
\value{
The function exports following object into global environment:
\itemize{
\item \code{q_alpha}: \eqn{\bm{q}(\bm{\alpha}^*)},
\item \code{A_const}: \eqn{\bar{\mathbf{A}}},
\item \code{A_alpha}: \eqn{\mathbf{A}(\bm{\alpha}^*)},
\item \code{gamma_alpha}: \eqn{\bm{\gamma}},
\item \code{data_end}: data frame of endogenous variables, \eqn{\bm{y}_t, t=0, 1, \cdots, T},
\item \code{data_exo}: data frame of exogenous variables, \eqn{x_t, t=0, 1, \cdots, T-1},
\item \code{quant_alpha}: data frame of conditional quantile, \eqn{Q_{\bm{\alpha}^*}(\bm{Y}_t|I_t), t=0, 1, \cdots, T}.
}
}
\description{
The function simulates a \eqn{n}-dimensional \eqn{\text{QVAR}(1)} process with one exogenous variable,
given sample size \eqn{T}.
}
\details{
The process is defined as
\deqn{
Q_{\bm{\alpha}}(\bm{Y}_t|I_t) = \bm{q}(\bm{\alpha}) + \mathbf{A}(\bm{\alpha}) \bm{Y}_{t-1} + \bm{\gamma} X_t,\ \bm{\alpha} \in (0, 1)^{n}, \\
\mathbf{A}(\bm{\alpha}) = \bar{\mathbf{A}} + 0.15( \bm{\alpha} - 0.5 \times \bm{1}_n ) \bm{1}_n^{\intercal},\\
\bm{Y}_{i0} \sim \mathcal{t}(2),\ q(\alpha_i) = F^{-1}(\alpha_i),\ \gamma_i \sim \mathcal{U}(-0.2, 0.2), i = 1, 2, \cdots, n,
}
where \eqn{F^{-1}(\cdot)} is the quantile function of \eqn{t(2)}.
}
\examples{
n <- 5
alpha <- rep(0.95, n)
bayesQVAR::simuData(n, alpha, 150)
}
